Release version 1.1.0
This is technically a breaking change in that it breaks exhaustive
switching over the ConnectionStatus enum. However given that we're very
early days on v1, we've decided to consider this an "obvious mistake"
bug and address it across all platforms. Internal discussion in [1].

Changelog text taken from JS at cb1bdcd.

[1] https://ably-real-time.slack.com/archives/C02NY1VT3LY/p1761576788131929